requirements.txt
torch==2.0.1
numpy==1.24.4
PyYAML==6.0.2
tqdm==4.66.4
gym==0.26.2
gym-notices==0.0.8
matplotlib==3.7.3